requirements.txt
flask
aiohttp
python-dotenv
google-generativeai
langchain
langchain-google-genai
gradio-client
